+++ src/main/java/chess/domain/piece/multiple/MultiplePiece.java
@@ -4,11 +4,12 @@ import chess.domain.ChessBoard;
 import chess.domain.Color;
 import chess.domain.Position;
 import chess.domain.direction.Direction;
+import chess.domain.piece.AbstractPiece;
 import chess.domain.piece.Piece;
 import java.util.Collections;
 import java.util.List;
 
-public abstract class MultiplePiece extends Piece {
+public abstract class MultiplePiece extends AbstractPiece {
 
     private final List<Direction> moveDirections;
