+++ src/test/java/chess/domain/piece/multiple/BishopTest.java
@@ -8,7 +8,6 @@ import chess.domain.ChessBoard;
 import chess.domain.Color;
 import chess.domain.Position;
 import chess.domain.piece.Piece;
-import chess.domain.piece.multiple.Bishop;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
