+++ src/test/java/chess/domain/piece/single/KnightTest.java
@@ -7,7 +7,6 @@ import chess.domain.ChessBoard;
 import chess.domain.Color;
 import chess.domain.Position;
 import chess.domain.piece.Piece;
-import chess.domain.piece.single.Knight;
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
